Leave body param validation to external lib
