+++ connexion/operation.py
@@ -15,7 +15,6 @@ from copy import deepcopy
 import functools
 import logging
 
-import jsonschema
 from jsonschema import ValidationError
 
 from .decorators import validation
@@ -169,28 +168,12 @@ class Operation(SecureOperation):
         self.endpoint_name = flaskify_endpoint(self.operation_id)
         self.__undecorated_function = resolution.function
 
-        for param in self.parameters:
-            if param['in'] == 'body' and 'default' in param:
-                self.default_body = param
-                break
-        else:
-            self.default_body = None
-
         self.validate_defaults()
 
     def validate_defaults(self):
         for param in self.parameters:
             try:
-                if param['in'] == 'body' and 'default' in param:
-                    param = param.copy()
-                    if 'required' in param:
-                        del param['required']
-                    if param['type'] == 'object':
-                        jsonschema.validate(param['default'], self.body_schema,
-                                            format_checker=jsonschema.draft4_format_checker)
-                    else:
-                        jsonschema.validate(param['default'], param, format_checker=jsonschema.draft4_format_checker)
-                elif param['in'] == 'query' and 'default' in param:
+                if param['in'] == 'query' and 'default' in param:
                     validation.validate_type(param, param['default'], 'query', param['name'])
             except (TypeValidationError, ValidationError):
                 raise InvalidSpecification('The parameter \'{param_name}\' has a default value which is not of'
@@ -386,7 +369,7 @@ class Operation(SecureOperation):
         if self.parameters:
             yield ParameterValidator(self.parameters)
         if self.body_schema:
-            yield RequestBodyValidator(self.body_schema, self.default_body is not None)
+            yield RequestBodyValidator(self.body_schema)
 
     @property
     def __response_validation_decorator(self):
