+++ tests/test_validation.py
@@ -1,5 +1,4 @@
 import json
-import pytest
 # we are using "mock" module here for Py 2.7 support
 from mock import MagicMock
